:sparkles: Add methods SetColumnOrder and AddRows to Table
